Clean-up
Removed EndpointHandlerBase - this class is not necessary anymore,
because all async handlers use AsyncHandlerBase.

Removed old SOAP handlers (ServiceStack uses async SOAP handlers now).

Removed formats from the EndpointAttributes enum. The user can already
inspect the Content-Type of the request from IHttpRequest, so there's no
need for this extra abstraction.

Additionally EndpointAttributes.LocalSubnet does not exist anymore since
the overhead to check if the request is from the local subnet is too
big. ServiceStack now uses IHttpRequest.IsLocal to determine if the
request is from localhost or external (if a user wants to know if the
request is from the local subnet, they should handle this information
themselves).

Removed ServiceAttribute and the corresponding tests - this
functionality will be replaced with a filter attribute.